% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{data}
\name{model}
\alias{Optimization.Model}
\alias{model}
\title{Create a New Optimization Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model()

Optimization.Model
}
\description{
A constructor of an object of type \code{"Optimization.Model"}.
}
\details{
Each document has the following methods:
}
\examples{
m <- model()
}
\keyword{datasets}

