% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize.R
\name{optimize}
\alias{optimize}
\title{optimize}
\usage{
optimize(model, solver = "auto", data = list(), ...)
}
\arguments{
\item{model}{an object of type \code{"Optimization.Model"}.}

\item{solver}{a character string giving the name of the solver.}

\item{data}{a list giving the data used in the the model.}

\item{...}{optional arguments.}
}
\description{
optimize a given model.
}

