% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\docType{data}
\name{ROML}
\alias{ROML}
\title{ROML - R Optimization Modeling Language}
\format{An object of class \code{environment} of length 4.}
\usage{
ROML
}
\description{
ROML - R Optimization Modeling Language
}
\keyword{datasets}

