% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\name{ROML_get_constraint_function}
\alias{ROML_get_constraint_function}
\title{Get Objective Transformation Funtion}
\usage{
ROML_get_constraint_function(x = NULL)
}
\arguments{
\item{x}{a character string giving the name of the funtion}
}
\description{
TODO
}
\details{
TODO
}
\examples{
ROML_get_constraint_function("sum")
}

