% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\name{ROML_get_objective_data_update_function}
\alias{ROML_get_objective_data_update_function}
\title{Get Objective Variable Generator Funtion}
\usage{
ROML_get_objective_data_update_function(x = NULL)
}
\arguments{
\item{x}{a character string giving the name of the funtion}
}
\description{
TODO
}
\details{
TODO
}

