% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\name{ROML_get_objective_function}
\alias{ROML_get_objective_function}
\title{Get Objective Transformation Function}
\usage{
ROML_get_objective_function(x = NULL)
}
\arguments{
\item{x}{a character string giving the name of the function}
}
\description{
Get an objective function by name.
}

