% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\name{ROML_register_constraint_function}
\alias{ROML_register_constraint_function}
\title{Register Special Function}
\usage{
ROML_register_constraint_function(name, method)
}
\arguments{
\item{name}{a character string giving the name of the function}

\item{method}{a method to be registered}
}

