% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROML.R
\name{ROML_register_objective_data_update_function}
\alias{ROML_register_objective_data_update_function}
\title{Register a Generator Function}
\usage{
ROML_register_objective_data_update_function(name, method)
}
\arguments{
\item{name}{a character string giving the name of the function}

\item{method}{a method to be registered}
}
\description{
TODO
}

